e2e: extract shared audio helpers and drop dev-only override
- Lift ensureNoPasswordViaAPI and waitForAudioStream into helpers.ts (the
  audio spec was inlining both, the latter as a copy of waitForVideoStream).
- ra-audio.spec.ts shrinks from 78 to 55 lines.
- Remove the JETKVM_AUDIO_DEVICE override in the remote agent: it fabricated
  an AudioDeviceInfo with is_jetkvm=true regardless of what device the env
  var pointed at, silently lying to the spec's assertion. Audio device
  discovery via aplay + /proc/asound/.../usbid is reliable; if no JetKVM
  device is present the spec already skips.
